% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_susr_data.R
\name{fetch_susr_data}
\alias{fetch_susr_data}
\title{Retrieve data from SUSR (Slovak Statistical Office) open data API as data frames}
\usage{
fetch_susr_data(
  params,
  lang = "en",
  base_url = "https://data.statistics.sk/api/v2/dataset"
)
}
\arguments{
\item{params}{A list structured in pairs:
\enumerate{
\item Table code (character string), e.g. "np3106rs"
\item Dimension specification (list/vector)
}}

\item{lang}{The language code, "en" or "sk". Defaults to "en".}

\item{base_url}{The base SUSR dataset endpoint. Defaults to
"https://data.statistics.sk/api/v2/dataset".}
}
\value{
A named list of data frames, keyed by the table code
}
\description{
This function constructs and calls the dataset endpoint for one or more tables,
including their dimension selections, and returns each as a data frame (tibble).
}
\details{
This function is designed to handle dimension queries that might include
special keywords (e.g., \code{"all"}, \code{"last5"}, or region shortcuts like \code{"Districts"}).

\strong{Error & Warning Handling}
We use \code{tryCatch()} blocks around both network calls and JSON parsing.
While we know some specific issues that could occur (like an invalid URL, or
invalid JSON structure), there may be other rare/unexpected problems.
If an error occurs, we issue a warning and store \code{NULL} for that table.

\strong{Dimension Count Check}
Before constructing the URL, we call \code{\link[=susr_tables]{susr_tables()}} (in "long" format) to retrieve
a list of the dimension codes for the table. We then check if the user-supplied
dimension specs have the same length as the \strong{number of distinct dimension_code}.
If not, we warn the user. (This ensures at least the total dimension count is valid.)
}
